+++ library/core/src/ffi/c_char.md
