+++ library/core/src/ffi/c_double.md
