+++ library/core/src/ffi/c_ushort.md
